Few bug fixing in calib utilities and cleanup of the code
