+++ share/goddess_134Xe_runs52_to_146.config
@@ -463,22 +463,22 @@ enCal n 3 0 1
 superX3 3015-17 D4-E1 16 153 16 309
 thresh p 0 0 0 0 0 0 0 0 
 thresh n 0 0 0 0 
-enCal p 0 0 1
-enCal p 1 0 1
-enCal resStrip 0 0 1
-posCal resStrip 0 0 0
-enCal p 2 0 1
-enCal p 3 0 1
-enCal resStrip 1 0 1
-posCal resStrip 1 0 0
-enCal p 4 0 1
-enCal p 5 0 1
-enCal resStrip 2 0 1
-posCal resStrip 2 0 0
-enCal p 6 0 1
-enCal p 7 0 1
-enCal resStrip 3 0 1
-posCal resStrip 3 0 0
+enCal p 0 3949.52 0.960691
+enCal p 1 27.0866 1
+enCal resStrip 0 0 4.3912e-05
+posCal resStrip 0 -0.47 0.4473
+enCal p 2 -6322.05 0.928796
+enCal p 3 -1680.61 1
+enCal resStrip 1 0 3.91121e-05
+posCal resStrip 1 -0.440766 0.340131
+enCal p 4 2983.85 1
+enCal p 5 1048.65 1.13271
+enCal resStrip 2 0 3.57877e-05
+posCal resStrip 2 -0.4089 0.45578
+enCal p 6 24.7516 1
+enCal p 7 -3596.84 0.844134
+enCal resStrip 3 0 3.92376e-05
+posCal resStrip 3 -0.462195 0.345906
 enCal n 0 0 1
 enCal n 1 0 1
 enCal n 2 0 1
@@ -486,22 +486,22 @@ enCal n 3 0 1
 superX3 2951-5 D5-E1 16 161 16 353
 thresh p 0 0 0 0 0 0 0 0 
 thresh n 0 0 0 0 
-enCal p 0 0 1
-enCal p 1 0 1
-enCal resStrip 0 0 1
-posCal resStrip 0 0 0
-enCal p 2 0 1
-enCal p 3 0 1
-enCal resStrip 1 0 1
-posCal resStrip 1 0 0
-enCal p 4 0 1
-enCal p 5 0 1
-enCal resStrip 2 0 1
-posCal resStrip 2 0 0
-enCal p 6 0 1
-enCal p 7 0 1
-enCal resStrip 3 0 1
-posCal resStrip 3 0 0
+enCal p 0 4404.85 0.956998
+enCal p 1 3156.86 1
+enCal resStrip 0 0 3.98579e-05
+posCal resStrip 0 -0.423723 0.427731
+enCal p 2 16546.8 0.962101
+enCal p 3 10160.8 1
+enCal resStrip 1 0 4.82776e-05
+posCal resStrip 1 -0.65222 0.44734
+enCal p 4 9463.25 1
+enCal p 5 14057.1 0.968949
+enCal resStrip 2 0 4.38186e-05
+posCal resStrip 2 -0.489828 0.391696
+enCal p 6 4542.08 1
+enCal p 7 6284.89 0.921912
+enCal resStrip 3 0 4.25665e-05
+posCal resStrip 3 -0.471928 0.374844
 enCal n 0 0 1
 enCal n 1 0 1
 enCal n 2 0 1
@@ -530,8 +530,8 @@ enCal n 1 0 1
 enCal n 2 0 1
 enCal n 3 0 1
 superX3 2951-7 D7-E1 19 121 19 305
-thresh p 0 0 0 0 0 0 0 0
-thresh n 0 0 0 0
+thresh p 0 0 0 0 0 0 0 0 
+thresh n 0 0 0 0 
 enCal p 0 94.144 1.07913
 enCal p 1 12.1603 1
 enCal resStrip 0 0 0.00268409
@@ -555,22 +555,22 @@ enCal n 3 0 1
 superX3 2951-8 D8-E1 19 129 19 309
 thresh p 0 0 0 0 0 0 0 0 
 thresh n 0 0 0 0 
-enCal p 0 0 1
-enCal p 1 0 1
-enCal resStrip 0 0 1
-posCal resStrip 0 0 0
-enCal p 2 0 1
-enCal p 3 0 1
-enCal resStrip 1 0 1
-posCal resStrip 1 0 0
-enCal p 4 0 1
-enCal p 5 0 1
-enCal resStrip 2 0 1
-posCal resStrip 2 0 0
-enCal p 6 0 1
-enCal p 7 0 1
-enCal resStrip 3 0 1
-posCal resStrip 3 0 0
+enCal p 0 106.861 0.895485
+enCal p 1 72.2779 1
+enCal resStrip 0 0 0.00362796
+posCal resStrip 0 -0.861076 0.860453
+enCal p 2 100.178 1.10649
+enCal p 3 59.7876 1
+enCal resStrip 1 0 0.00307158
+posCal resStrip 1 -0.625193 0.9082
+enCal p 4 56.2854 1
+enCal p 5 95.2358 0.989055
+enCal resStrip 2 0 0.00318978
+posCal resStrip 2 -0.673584 0.90838
+enCal p 6 58.3445 1
+enCal p 7 84.8319 1.20992
+enCal resStrip 3 0 0.00292313
+posCal resStrip 3 -0.542943 1.06574
 enCal n 0 0 1
 enCal n 1 0 1
 enCal n 2 0 1
