+++ toolbox/include/GoddessToolbox.h
@@ -43,6 +43,8 @@ using std::clog;
 using std::endl;
 using std::size_t;
 
+const char esc ( 27 );
+
 //_______________________________________________________________________________________________________________________________________________//
 //_______________________________________________________________________________________________________________________________________________//
 //_______________________________________________________________________________________________________________________________________________//
